+++ types/custom-event.ts
@@ -6,17 +6,3 @@ export type Selection = {
   secId: string;
   step: string;
 };
-
-export function customEventCallStackUpdate() {
-  window.dispatchEvent(new CustomEvent(CUSTOM_EVENT_CALLSTACK_UPDATE, {}));
-}
-
-export function customEventSelection(selection: Selection) {
-  window.dispatchEvent(
-    new CustomEvent(CUSTOM_EVENT_SELECTION, { detail: selection }),
-  );
-}
-
-export function customEventSDOSelection() {
-  window.dispatchEvent(new CustomEvent(CUSTOM_EVENT_SDO_SELECTION, {}));
-}
